Move Auth0 Into Core
