+++ apps/zui/src/domain/lake/handlers.ts
@@ -2,7 +2,7 @@ import {Client} from "@brimdata/zed-js"
 import {createHandler} from "src/core/handlers"
 import {LakeModel} from "src/js/models/lake"
 import Current from "src/js/state/Current"
-import {validateToken} from "src/js/auth0/utils"
+import {validateToken} from "src/core/auth0/utils"
 import {getAuthCredentials} from "src/js/flows/lake/getAuthCredentials"
 import Lakes from "src/js/state/Lakes"
 import LakeStatuses from "src/js/state/LakeStatuses"
