+++ apps/zui/src/js/flows/lake/getAuth0.ts
@@ -1,5 +1,5 @@
 import {LakeModel} from "src/js/models/lake"
-import Auth0Client from "../../auth0"
+import Auth0Client from "../../../core/auth0"
 import Current from "../../state/Current"
 import {Thunk} from "../../state/types"
