+++ apps/zui/src/js/flows/lake/update-status.ts
@@ -1,7 +1,7 @@
 import {Client} from "@brimdata/zed-js"
 import {syncPoolsData} from "src/app/core/pools/sync-pools-data"
 import lake from "src/js/models/lake"
-import {validateToken} from "../../auth0/utils"
+import {validateToken} from "../../../core/auth0/utils"
 import Lakes from "../../state/Lakes"
 import LakeStatuses from "../../state/LakeStatuses"
 import {getAuthCredentials} from "./getAuthCredentials"
